Fixed lazy route files and some e to e tests
