+++ web/src/routes/_auth/settings/playback.tsx
@@ -1,4 +1,4 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useId, useState, useTransition } from "react";
 import type { FormEvent, ReactNode } from "react";
@@ -52,7 +52,7 @@ import type {
   UpdatePlaybackSettingsRequest,
 } from "@/types";
 
-export const Route = createLazyFileRoute("/_auth/settings/playback")({
+export const Route = createFileRoute("/_auth/settings/playback")({
   component: PlaybackSettings,
 });
