+++ web/src/routes/_auth/watch-rooms/$id.tsx
@@ -1,5 +1,9 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { watchRoomQueryOpts } from "@/lib/query-opts";
+import {
+  WatchRoomPage,
+  WatchRoomUnavailable,
+} from "@/components/watch-room/WatchRoomPage";
 
 export const Route = createFileRoute("/_auth/watch-rooms/$id")({
   params: {
@@ -14,6 +18,7 @@ export const Route = createFileRoute("/_auth/watch-rooms/$id")({
     if (id === null) return;
     await context.queryClient.ensureQueryData(watchRoomQueryOpts(id));
   },
+  component: WatchRoomRoute,
 });
 
 function parseWatchRoomId(value: string) {
@@ -23,4 +28,20 @@ function parseWatchRoomId(value: string) {
   }
 
   return parsed;
+}
+
+function WatchRoomRoute() {
+  const { id } = Route.useParams();
+  const navigate = Route.useNavigate();
+
+  if (id === null) {
+    return (
+      <WatchRoomUnavailable
+        message="This watch room link is invalid."
+        onBackHome={() => navigate({ to: "/" })}
+      />
+    );
+  }
+
+  return <WatchRoomPage roomId={id} />;
 }
